Switch back to logging through a queue (of the non-blocking variety)
Log messages were originally passed around using a queue, but it was
slow because it was used in a blocking fashion. That isn't necessary,
and using a queue prevents race conditions around log mutation (which
weren't present but could be triggered if a caller of server.log() does
something nasty).