Communicate shutdown to the server thread via pipe
Using a pipe (instead of a threading.Event) allows us to make the
server's select() loop fully blocking (it now waits for reads from the
socket and the shutdown pipe).

This provides a huge speedup as we're no longer looping in Python
while waiting for requests (which was the case when select() had
a timeout specified).